requirements.txt
djangorestframework
django-editorjs-fields
django-editorjs
django-environ
django-mptt
psycopg2
psycopg2-binary
Pillow
